Changed mobile ads to be on top of page for all pages except front page.
Ads on mobile were being placed at the bottom of each article. To fix this, I removed the ad from each article build and instead included it in the PageWrapper. So, on all pages except the front page the ad will be on top for mobile users. Tablet and desktop are not affected by this commit.